feat: update badge counter logic
- Moved logic to util file.
- Now the badge counter is synced with the filter logic for notifications.
